Micro benchmarks tidy up
* Added switcher so we can add more benchmarks
* Added a benchmark comparing pub wait-until-sent option
